Fixes overflow problem when sampleName_trackname is too long
The variable sampleName_trackname in MakeTrackhub.py needs to be
106 characters long or less, else a buffer overflow error occurs.

This was happening due to long cegsvectors related names. This
fix shortens those names.

Also included is a fix to only copy the GC percentage file when
hub type is CEGS. For other hub types the copy fails, which is
not a problem but which generated an error message.