Finished modification of sbatch structure
Also added a file:  call_one_sample.bash
It shows how to make a simple command line call to the main code.

If this code is to be downloaded to a new location, then line
numbers 4 and 7 of call_launch_bam_intersect.bash should be modified
to reflect where the downloaded code exists, and where the output
will be dumped.