Add more pipes and change some variable names
fixes #72

except for point 1. I am hesitant to place files in TMPDIR in the
submit_bamintersect.sh script, because TMPDIR=/tmp in that script.

I notice that TMPDIR is used elsewhere in the script now, because
I used to have a TMPDIR=mktemp line in the script. That's missing now.