Makes NM=0 and ReqFullyAligned the defaults for calls to bamintersect.sh
Also changes an error code in bamintersect.py to zero.  Previously, when
there were no mapped reads for a chromosome in the first bam file, an
error was returned. But it turns out that sometimes a LP chromosome has
mapped reads, but the LP backbone does not. An error returned from
bamintersect.py would generate a pipefail error in bamintersect.sh, so
we avoid this.