Permit tuning mapping stringency (#91)
* Read the NM tag and require reads to be fully aligned by default. fixes#90

* Also changes an error code in bamintersect.py to zero.  Previously, when
there were no mapped reads for a chromosome in the first bam file, an
error was returned. But it turns out that sometimes a LP chromosome has
mapped reads, but the LP backbone does not. An error returned from
bamintersect.py would generate a pipefail error in bamintersect.sh, so
we avoid this.