Simplify HA_table.sh
The homology arm coordinates are zero-based, so the code has been changed to use the "-L" option in the calls to samtools view. This causes samtools to use a bed file for extracting ranges.

Also, the 5p and 3p HA arms are no longer considered separately. One bed file is now generated for the HAs, instead of two.