Change hub deployment method (#173)
* Change hub deployment method - fixes #170 

Now building the hub in a tmp directory, hubchecking it there, and only then
deleting the old hub and replacing it with the new. Also, since we are now
using relative paths in our track definitions, we no longer need to pass the
hub's URL (and associated login/password) to update_tracks.bash

The usage comments in update_tracks.bash have also been updated to reflect the changes.