Select lines from samplesforTrackhub output via mappedgenome, not ann… (#195)
* Select lines from samplesforTrackhub output via mappedgenome, not annotationgenonome.

Also, do not search for exact matches.  The lines are selected by genome, which
is normally one of hg38, rn6, mm10, etc.  Just regex the Mapped_genome contents
for the correct genome.  If two genomes appear in the field (as in hg38_full_wuhCor1),
then it is OK to have the line appear as input to both trackhub files.

fixes #192

Co-authored-by: mattmaurano <44711377+mattmaurano@users.noreply.github.com>