Limit tree traversal, and add pipefail
Allow subdirectories to exist in the directories which contain assembly bed files.
The code will only look for bed files in the top level directory, allowing other
bed files to exist in the subdirectories.  These will not be used in hub construction.

Also add "set -eu -o pipefail" functionality to the bash scripts.