Call merge_bamintersect with positional arguements rather than exporting variables
Similar to previous update for sort_bamintersect and bamintersect.

I left the defaults related to no_bed, no_countstable, and noReqFullyAligned.
They set values for the make_bed, make_table, and ReqFullyAligned variables.

However, if submit_bamintersect is called with --no_bed, then
make_bed="", and while this was OK when using export, this would
now cause an unwanted shift in the positional variable assignments
in multiple places. Want me to fix this, or remove the use of
no_bed and no_countstable all together?