Replace mlr with sort to fix counts_table.sh
At line 155 in the counts_table.sh pipeline, there are no headers in the output from annotateNearestGeneName. This headerless output is being passed as input to mlr.
In the particular case of BS04412A, for each +/- strand combination there are only zero or one lines coming out from annotateNearestGeneName. So that one line is
interpreted by mlr as a header, and no output at all is passed from mlr to awk. This eventually results in an empty counts.txt file.

This fix replaces mlr with a call to standard sort.