SpeedUp AdjacencyDeDup.py
I was investigating the issue freezing Raquel Ordonez processing and besides some other changes with minor speed improvement, the main issue I detect is the inclusion check on `get_connected_components_adjacency` which uses a list instead of a set that is much faster. On my tests, the execution that ran overnight was able to complete in a couple of minutes.