Updated pipeline to include Delly snp calls
This commit runs delly and generates a bcf and vcf file per genome.

If delly returns a non-zero exit code then these files are not
generated, and a generic error message is added to the log files.

Delly is now imported in submit.sh

Amendment: Removed hard coded url that gitguard flagged