Pipeline creates a .bb file for delly output
Delly is now run in callsnpsMerge and is not parallelized by chromosome

ParseSamtoolsGenotypesToBedFiles.pl was updated to parse delly format
strings. Extra steps were added to extract the end coordinate from the
INFO values.

Fix: Delly outputs genome quality score which are > 1000 resulting in an
error when converting the bed files to big bed format. To bypass this
the values are clamped to 1000.