Delly and bcftools vcfs are merged for bigbed
Delly vcf output is concatenated with bcftools vcf output.
Both original files are written to the output directory, in the future
these files can most likely just be written to TMP and only the merged
is stored.

I added an import for R which was missing from the submit.sh script