Concatenated bigbed
Delly and bcftools outputs are written to TMP and the joined VCF is
written to the output directory.

Updated the variables that are used in the bed file by Delly
  - $readDepth = $PE
  - $PL = 'NA'

Added OPENBlas as a dependency which was missing.

Added a FIXME to callsnpsByChrom as I think these files should be
written to TMPDIR rather that the output as they are deleted during the
merge step.