Fix/starchcat double free (#280)
Fixes #275

Changes:
Updates the modules to use the latest version of bedops (2.4.42)

Why:
The previous version of bedops (2.4.41) had a bug in the starchcat script that would result in a double-free error in particular circumstances. This would cause the pipeline to fail as the analysis job would fail with an error.

Testing:
I reran the pipeline on the sample that had the error and checked that it completed successfully.